moves to contextualized folder
